Add a data.json file
We want to publish agreements under the Digital Economies Act on data.gov.uk.
To do this we need to publish some metadata. We've chosen the data.json format
because it's easier to produce than DCAT while still letting us include all the
fields needed to publish on data.gov.uk:
https://guidance.data.gov.uk/publish_and_manage_data/harvest_or_add_data/